Add files via upload
The artillery function allows the player to target a 2x2 area. When iterating through this area, if it founds 'S' (which means a ship was there) it replaces it by '*' to mark the hit and return true  .If it founds  '~' (indicating no ship in that location) it replaces it by 'o' to mark the miss.