Ship placing update
Major changes:
Added Shipboard (type int) to PlayerStatus struct, which is a copy of the original board (type char) where:
Every (~) is a 0.
Every S (marker for ship) is its corresponding size, ex. a ship of size 5 will have: 5 S's in a row on Board, 5 5s in a row on ShipBoard. 
This helps with later ship sinking logic, checking for win, checking for attacks...
Also added Ship placing method along with method that displays which ships are available, takes input coordinates and orientation and then calls placeShip, that places the ship on both the board and the shipboard.
Added initialization for both boards.
Also added displayBoard into the stunt:
Display board is the point of view of the player's opponent's boards, only showing hits and misses in easy mode, hits only in hard mode. This allows for more dynamic gameplay as the player can see previous moves and outcomes, depending on difficulty chosen.