Abilities update
Started with abilities and gameplay aspect of the game:
first we added to the structure some fields that helps us such as counter for each abilities and bool flags for conditional abilities, ex. Torpedo, smokeScreens..
We also implemented displayAvailableMoves, which displays only available moves, takes input (also checks validity of input/coordinates), calls methods involved and updates what needs to be updated respectively. We started with implementing the attack method, and decided as a group that it should be the main focus to reuse it in different methods, as seen with the code above in the torpedo method, which ultimately just calls attack method on a single row/column depending on input. We also took into consideration edge cases such as row 10, since it doesn't operate as the other 9 rows, since its a 2 char, hence to solve the problem, we used a char[] instead of taking input as int, to make it easier to deal with such edge cases.