Added artillery and check Win
Artillery is an attack move, it targets a 2x2 area. The checkWin method checks for the presence of an 'S' on the board, which indicates that there is still a ship that hasn't been sunk. If it finds an 'S', it returns false. If no 'S' is found, it returns true, meaning  that all ships have been sunk and the game is won.