ship sinking logic + end phase 1
The checkSunkShips method is responsible for determining the number of ships sunk at the end of each round. It verifies that all parts of every ship are completely sunk before marking them as such. This method is also utilized by other functions, such as torpedo and artillery.

For example, in the case of the torpedo function, a counter in the struct is initialized to 0. When three ships have been sunk, the counter increments to 1, allowing the canUseTorpedo flag to become true for the next round only. This behavior is enforced by an if condition requiring counter < 1.

Additionally, three separate counters—shipOneCounter, shipTwoCounter, and shipThreeCounter—are added to the struct and initialized to 0. These counters are updated within checkSunkShips. When the first ship is sunk, shipOneCounter is set to 1, enabling the use of artillery for the next round only. Similarly, when the second and third ships are sunk, shipTwoCounter and shipThreeCounter are set to 1, respectively, each unlocking artillery for a single subsequent round. This behavior is enforced by the if statements which requires that all 3 counters< 1

These counters ensure that torpedo and artillery functions are available only under the correct conditions and for the appropriate rounds.