Medium Difficulty bot implementation + a lot of debugging
We implemented the medium difficulty bot:
This bot is an improved version of the Easy bot previously committed; they both start by firing at random cells, but this difficulty ensures that once a hit is detected, the next turn adjacent cells to that hit cell will be targeted. This was done by adding a botStatus structure, that keeps count of the lastHit x and y coordinates, and a boolean flag that indicates if last turn had a hit or not. This bot also utilizes abilities more efficiently than easyBot; adding the "Targeting" mode ensures a more dynamic gameplay vs the bot. Radar sweeping and finding a ship will mark that coordinate as a coordinate of focus for next turn to be attacked, same for artillery and torpedo, which both have 90% of getting called once obtained through sinking ships, while the non violent but strategic abilities such as radar sweep and smokescreen have a 20% chance of being used, due to the fact that once obtained in arsenal, they don't go away next turn like artillery and torpedo. 
Also we fixed bugs in radar sweep method where we didn't check for overlap before, which caused the bot to sweep one end of the board and getting a true flag for ships on the other side. We still have to implement HardBot which will be an even smarter version of this bot and to add a difficulty setter in the main for the variable diff (which we already have the input correctly from the user). And finally we have to add the win statements at the end.