Hard Bit update + End of phase 2 AND PROJECT!
We finally added hardbot implementation, this bot is more efficient in hitting ships, using abilities and strategically more sound than the other 2 versions, which relied more on random factor. This one utilizes some functionality from the other 2, and improves on t, especially targeting mode, which can guarantee a 35% chance of hitting a ship.
This is the last step of the project, good work everyone!
Onto report and testing next.